--- src/app/labs/imagen/page.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-import '../../_components/_styles/labs.css';
-import Labslab from './components/Labslab.jsx';
-import Footer from './components/footer';
-
-const ImagenSite = () => {
-	return (
-		<div className="App">
-			<div className="container">
-			<div className='form-container'>
-			<Labslab />
-        <h2>IMAGEN HERE!</h2>
-			  </div>
-			  <Footer />
-		  </div> 
-    </div>
-	);
-};
-  
-export default ImagenSite;
