+++ src/app/labs/mint/page.tsx
@@ -1,14 +1,18 @@
-// src/app/mint/pages.tsx //TODO
-import React from 'react';
+// MINT - src/app/labs/mint/page.tsx // REVIEW
+'use client'
+import React, { useEffect, useState } from "react";
 import "../_components/_styles/labs.css";
-import HomeBar from '../_components/_site/Homebar';
-import Footer from '../_components/_site/Footer';
-import { Account } from '../_components/_web3/Account';
-import { Connect } from '../_components/_web3/Connect';
-import { Connected } from '../_components/_web3/Connected';
-import { MintNFT } from '../_components/_web3/MintNFT';
-import { NetworkSwitcher } from '../_components/_web3/NetworkSwitcher';
+import LabNav from '../../_components/_labs/LabNav';
+import LabFooter from '../../_components/_labs/LabFooter';
+import { Button, TextField, Grid, Box } from '@mui/material';
 
+import Image from 'next/image';
+import { Account } from '../../_components/_web3/Account';
+import { Connect } from '../../_components/_web3/Connect';
+import { Connected } from '../../_components/_web3/Connected';
+import { MintNFT } from '../../_components/_web3/MintNFT';
+import { NetworkSwitcher } from '../../_components/_web3/NetworkSwitcher';
+import mintABI from '../../_components/_labs/_utils/mintABI.json';
 
 const OPENSEA_URL = "https://testnets.opensea.io/assets/";
 const NETSCAN_URL = "https://mumbai.polygonscan.com/tx/";
@@ -18,14 +22,19 @@ const MetaMaskURL = "https://metamask.io/download/";
 const CBWalletURL = "https://chrome.google.com/webstore/detail/coinbase-wallet-extension/hnfanknocfeofbddgcijnmhnfnkdnaad/";
 const MaticURL = "https://wallet.matic.network/";
 
+const CONTRACT_ADDRESS = "0x22FACFbf9dd893Ec8a0aF6d0764c9d30b27D5Bc3";
+const contractABI = mintABI.abi;
+const TOTAL_MINT_COUNT = 10;
+
 const MintSite: React.FC = () => {
   return (
     <div className="App">
       <div className="container">
         <div className="header">
-          <HomeBar />
+          <LabNav />  
         </div>
         <div className="middle-row">
+        <h2>MINT PORTAL</h2>
         <Connect />
           <Connected>
             <Account />
@@ -36,7 +45,7 @@ const MintSite: React.FC = () => {
             <br />
           </Connected>
         </div>
-        <Footer />
+        <LabFooter />
       </div>
     </div>
   );
