+++ src/app/projects/blogs/_components/BlogPanel.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import styles from "../styles/Blog.module.css";
+import Image from 'next/image';
 
 export default function BlogPanel({ blog }) {
   const { id, name, description, image_url } = blog;
@@ -7,7 +8,7 @@ export default function BlogPanel({ blog }) {
   return (
     <div className={styles.blog_container}>
       <div >
-        <img className={styles.blog_image} src={image_url} alt="" />
+        <Image className={styles.blog_image} src={image_url} alt="" />
       </div>
 
       <div className={styles.blog_details}>
