+++ src/app/projects/clients/_comp/buttonz.tsx
@@ -1,7 +1,7 @@
 import PropTypes from "prop-types";
 import React from "react";
-import { RocketLaunch } from "./RocketLaunch";
-import "../styles/style.css";
+import { RocketLaunch } from "./rocket";
+import "./_styles/client.css";
 
 interface Props {
   hasLeftIcon: boolean;
